Move node objects to respective file
