+++ dds/src/implementation/actors/any_data_writer_listener.rs
@@ -1,5 +1,5 @@
 use crate::{
-    implementation::dds::nodes::DataWriterNode,
+    implementation::dds::dds_data_writer::DataWriterNode,
     infrastructure::status::{
         LivelinessLostStatus, OfferedDeadlineMissedStatus, OfferedIncompatibleQosStatus,
         PublicationMatchedStatus,
