+++ dds/src/implementation/actors/data_writer_listener_actor.rs
@@ -1,7 +1,7 @@
 use dust_dds_derive::actor_interface;
 
 use crate::{
-    implementation::dds::nodes::DataWriterNode,
+    implementation::dds::dds_data_writer::DataWriterNode,
     infrastructure::status::{OfferedIncompatibleQosStatus, PublicationMatchedStatus},
 };
