+++ dds/src/implementation/dds/mod.rs
@@ -1,2 +1,6 @@
-pub mod nodes;
+pub mod dds_data_reader;
 pub mod dds_domain_participant;
+pub mod dds_publisher;
+pub mod dds_subscriber;
+pub mod dds_topic;
+pub mod dds_data_writer;
